Updated to Ingress Controller
Turns out that the VirtualServer custom resource is incompatible with CertManager. The CertManager wants to start an Ingress object with the same hostname as Foundry. But because the VirtualServer owns that name, it can't launch and eventually the certificate from LetsEncrypt expires and can't be renewed. Confirmed that the team for NGINX is still working on the ACME challenge/response process. Moved everything to an Ingress object to handle things.